+++ R/form-controls.R
@@ -15,7 +15,7 @@ block_textarea <- function(
   input_id,
   value = "",
   placeholder = NULL,
-  rows = 4,
+  rows = 2,
   width = NULL,
   disabled = FALSE,
   class = NULL